+++ Goobi/src/org/goobi/beans/Process.java
@@ -879,9 +879,9 @@ public class Process implements Serializable, DatabaseObject, Comparable<Process
         double error2 = 0;
         double abgeschlossen2 = 0;
 
-        offen2 = (offen * 100) / (double) (offen + inBearbeitung + error + abgeschlossen);
-        error2 = (error * 100) / (double) (offen + inBearbeitung + error + abgeschlossen);
-        inBearbeitung2 = (inBearbeitung * 100) / (double) (offen + inBearbeitung + error + abgeschlossen);
+        offen2 = (offen * 100) / (offen + inBearbeitung + error + abgeschlossen);
+        error2 = (error * 100) / (offen + inBearbeitung + error + abgeschlossen);
+        inBearbeitung2 = (inBearbeitung * 100) / (offen + inBearbeitung + error + abgeschlossen);
         abgeschlossen2 = 100 - offen2 - inBearbeitung2 - error2;
         return abgeschlossen2;
     }
